Add -f flag to specify a field other than DP
If the specified field can be found, the threshold will be applied to that
field. If the specified field cannot be found, the filter will be applied to
the DP field instead. This is most useful when filtering on MIN_DP, since
MIN_DP is not defined for variants, but DP is.